requirements.txt
Flask==2.2.5
pymysql==1.0.3
python-dotenv==0.21.0
cryptography